final-pjt-back/requirements.txt
asgiref==3.4.1
autopep8==1.5.7
certifi==2021.10.8
charset-normalizer==2.0.7
Django==3.2.6
django-appconf==1.0.5
django-cors-headers==3.10.0
django-imagekit==4.1.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
idna==3.3
pilkit==2.0
Pillow==8.4.0
pycodestyle==2.7.0
PyJWT==1.7.1
python-decouple==3.5
pytz==2021.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
typing-extensions==4.0.0
urllib3==1.26.7
